+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+ - Add class `Netzmacht\ContaoFormBundle\Form\FormGenerator\UploadHandler` and service 
+   `netzmacht.contao_form.form_generator.upload_handler` to handle form uploads.
+
 ## [1.0.2] - 2019-06-20
 
 ### Fixed
@@ -17,5 +24,6 @@
  - Do not include invisible form fields
  - Fix choices group for multiple select fields
 
+[Unreleased]: https://github.com/netzmacht/contao-form-bundle/compare/1.0.2...dev-develop
 [1.0.1]: https://github.com/netzmacht/contao-form-bundle/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/netzmacht/contao-form-bundle/compare/1.0.0...1.0.1